ISSUE #478 add test case for IFC4, unify ifc4 and ifc2x3 code to reduce duplication
