+++ ifcUtils/repo_ifc_utils_Ifc2x3.h
@@ -16,107 +16,6 @@
 */
 
 #pragma once
-#include "repo_ifc_utils_globals.h"
-#include <ifcparse/Ifc2x3.h>
-#include <ifcparse/IfcParse.h>
-#include <ifcparse/IfcFile.h>
-
-//FIXME
-#include "../bouncer/src/repo/lib/datastructure/repo_structs.h"
-
-namespace repo {
-	namespace ifcUtility {
-		namespace Schema_Ifc2x3 {
-			class IFC_UTILS_API_EXPORT GeometryHandler {
-			public:
-				static bool retrieveGeometry(
-					const std::string &file,
-					std::vector < std::vector<double>> &allVertices,
-					std::vector<std::vector<repo_face_t>> &allFaces,
-					std::vector < std::vector<double>> &allNormals,
-					std::vector < std::vector<double>> &allUVs,
-					std::vector<std::string> &allIds,
-					std::vector<std::string> &allNames,
-					std::vector<std::string> &allMaterials,
-					std::unordered_map<std::string, repo_material_t> &matNameToMaterials,
-					std::vector<double>		&offset,
-					std::string              &errMsg);
-			};
-
-			class IFC_UTILS_API_EXPORT TreeParser
-			{
-			public:
-				bool missingEntities = false;
-
-				TreeParser() {}
-				~TreeParser() {}
-
-				static TransNode createTransformations(const std::string &filePath, bool &missingEntities);
-
-			protected:
-
-				static std::pair<std::string, std::string> processUnits(
-					const IfcUtil::IfcBaseClass *element);
-
-				static std::string getUnits(
-					const Ifc2x3::IfcDerivedUnitEnum::Value &unitType,
-					const std::unordered_map<std::string, std::string> &projectUnits
-				);
-
-				static std::string getUnits(
-					const Ifc2x3::IfcUnitEnum::Value &unitType,
-					const std::unordered_map<std::string, std::string> &projectUnits
-				);
-
-				static std::string getUnits(
-					const std::string &unitType,
-					const std::unordered_map<std::string, std::string> &projectUnits
-				);
-
-				static void generateClassificationInformation(
-					const Ifc2x3::IfcRelAssociatesClassification * &relCS,
-					std::unordered_map<std::string, std::string>    &metaValues
-
-				);
-
-				static TransNode createTransformationsRecursive(
-					IfcParse::IfcFile &ifcFile,
-					bool &missingEntities,
-					const IfcUtil::IfcBaseClass *element,
-					std::unordered_map<std::string, std::string>                               &metaValue,
-					std::unordered_map<std::string, std::string>                               &locationValue,
-					std::unordered_map<std::string, std::string>                               &projectUnits,
-					const std::set<int>													       &ancestorsID = std::set<int>(),
-					const std::string														   &metaPrefix = std::string()
-				);
-
-				static void determineActionsByElementType(
-					const IfcUtil::IfcBaseClass *element,
-					bool &missingEntities,
-					std::unordered_map<std::string, std::string>                  &metaValues,
-					std::unordered_map<std::string, std::string>                  &locationData,
-					std::unordered_map<std::string, std::string>                  &projectUnits,
-					bool                                                          &createElement,
-					bool                                                          &traverseChildren,
-					std::vector<IfcUtil::IfcBaseClass *>                          &extraChildren,
-					const std::string											  &metaPrefix,
-					std::string											          &childrenMetaPrefix);
-
-				static void setProjectUnits(
-					const Ifc2x3::IfcUnitAssignment* unitsAssignment,
-					std::unordered_map<std::string, std::string> &projectUnits);
-
-				static std::string constructMetadataLabel(
-					const std::string &label,
-					const std::string &prefix = "",
-					const std::string &units = ""
-				);
-
-				static std::string getValueAsString(
-					const Ifc2x3::IfcValue    *ifcValue,
-					std::string &unitType,
-					const std::unordered_map<std::string, std::string> &projectUnits);
-			};
-		}
-	}
-}
+#define SCHEMA_IN_FILE Ifc2x3
+#include "repo_ifc_utils_schema_template.h"
+#undef SCHEMA_IN_FILE
