+++ test/src/unit/repo_test_database_info.h
@@ -42,6 +42,7 @@ const static std::string missingNodesModel = "Wall.ifc";
 const static std::string texturedModel2 = "texturedPlane2.dae"; //With Texture
 const static std::string badExtensionFile = "cube.exe";
 const static std::string ifcModel = "duplex.ifc";
+const static std::string ifc4Model = "ifc4Test.ifc";
 const static std::string dgnModel = "sample.dgn";
 const static std::string dwgModel = "box.dwg";
 const static std::string dxfModel = "box.dxf";
